chat-memory/pyproject.toml
[project]
name = "chat-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "langchain-community==0.3.16",
    "langchain-core==0.3.33",
    "langchain-openai==0.3.3",
    "langsmith==0.3.4",
    "pydantic==2.9.2",
    "ipython>=8.37.0",
    "rich>=14.1.0",
]
